requirements.txt
faster-whisper>=1.1.0
openai>=1.12.0
ffmpeg-python>=0.2.0
pyyaml>=6.0
tqdm>=4.66.0
pysrt>=1.1.2
